implemented solver that outputs new schedules

+++ TrainSolution/src/trainUtils/Train.java
@@ -2,9 +2,9 @@ package trainUtils;
 
 public class Train {
 	public TrainType type;
-	public short aArr, bArr, cArr, aBoard, bBoard, cBoard;
+	public int aArr, bArr, cArr, aBoard, bBoard, cBoard;
 	
-	public Train(TrainType type, short aArr, short aBoard, short bArr, short bBoard, short cArr, short cBoard) {
+	public Train(TrainType type, int aArr, int aBoard, int bArr, int bBoard, int cArr, int cBoard) {
 		this.type = type;
 		this.aArr = aArr;
 		this.bArr = bArr;
